Bootstrapping Estimates for 'cond_indirect_effects'

Generate bootstrapping estimates to be used by cond_indirect_effects(), cond_indirect(), and indirect_effect().

Usage

do_boot(
  fit,
  R = 100,
  seed = NULL,
  parallel = TRUE,
  ncores = max(parallel::detectCores(logical = FALSE) - 1, 1),
  make_cluster_args = list(),
  progress = TRUE
)

Arguments

fit

Either (a) a list of lm class objects, or the output of lm2list() (i.e., a lm_list-class object), or (b) the output of lavaan::sem().

R

The number of bootstrap samples. Default is 100.

seed

The seed for the bootstrapping. Default is NULL and seed is not set.

parallel

Logical. Whether parallel processing will be used. Default is TRUE. If fit is a list of lm() outputs, parallel processing will not be used.

ncores

Integer. The number of CPU cores to use when parallel is TRUE. Default is the number of non-logical cores minus one (one minimum). Will raise an error if greater than the number of cores detected by parallel::detectCores(). If ncores is set, it will override make_cluster_args.

make_cluster_args

A named list of additional arguments to be passed to parallel::makeCluster(). For advanced users. See parallel::makeCluster() for details. Default is list().

progress

Logical. Display progress or not. Default is TRUE.

Value

A boot_out-class object that can be used for the boot_out

argument of cond_indirect() and cond_indirect_effects()

for forming bootstrapping confidence intervals.

Details

It do nonparametric bootstrapping to generate bootstrap estimates of the parameter estimates in a model fitted either by lavaan::sem() or by a sequence of calls to lm(). The stored estimates can then be used by cond_indirect_effects(), cond_indirect(), and indirect_effect() to form bootstrapping confidence intervals.

This approach removes the need to repeat bootstrapping in each call to cond_indirect() and cond_indirect_effects(). It also ensures that the same set of bootstrap samples is used in all subsequent analysis.

It determines the type of the fit object and then call lm2boot_out(), fit2boot_out(), or fit2boot_out_do_boot().

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

data(data_med_mod_ab1)
dat <- data_med_mod_ab1
lm_m <- lm(m ~ x*w + c1 + c2, dat)
lm_y <- lm(y ~ m*w + x + c1 + c2, dat)
lm_out <- lm2list(lm_m, lm_y)
# In real research, R should be 2000 or even 5000
lm_boot_out <- lm2boot_out(lm_out, R = 100, seed = 1234)
wlevels <- mod_levels(w = "w", fit = lm_out)
wlevels
#>                w
#> M+1.0SD 6.046455
#> Mean    4.990179
#> M-1.0SD 3.933902
out <- cond_indirect_effects(wlevels = wlevels,
                             x = "x",
                             y = "y",
                             m = "m",
                             fit = lm_out,
                             boot_ci = TRUE,
                             boot_out = lm_boot_out)
out
#> 
#> == Conditional indirect effects ==
#> 
#>  Path: x -> m -> y
#>  Conditioned on moderator(s): w
#>  Moderator(s) represented by: w
#> 
#>       [w]   (w)    ind  CI.lo CI.hi Sig    m~x   y~m
#> 1 M+1.0SD 6.046  0.248  0.030 0.458 Sig  0.342 0.725
#> 2 Mean    4.990  0.024 -0.079 0.171      0.063 0.375
#> 3 M-1.0SD 3.934 -0.006 -0.087 0.098     -0.216 0.026
#> 
#>  - [CI.lo to CI.hi] are 95.0% percentile confidence intervals by
#>    nonparametric bootstrapping with 100 samples.
#>  - The 'ind' column shows the indirect effects.
#>  - ‘m~x’,‘y~m’ is/are the path coefficient(s) along the path conditional
#>    on the moderators.
#>