Convert a 'lm' Output to a 'lavaan'-Like Object

It only convert information necessary for cond_indirect() to compute conditional effects or conditional indirect effects.

Usage

lm2ptable(outputs)

Arguments

outputs

A list of lm class objects.

Value

A list with the following arguments:

est: A data frame similar to the parameter estimates table of the output of lavaan::sem().

dat: A data frame with all the data in the regression models merged.

implied_stats: A list with two elements: cov is the implied covariance matrix and mean is the implied means. They are used to compute standardized effects or form levels of moderators.

Details

Users usually do not need to call this function to use cond_indirect() and cond_indirect_effects(). These two functions will do the conversion internally if necessary.

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

data(data_serial)
dat <- data_serial
lm_m1 <- lm(m1 ~ x + c1 + c2, dat)
lm_m2 <- lm(m2 ~ m1 + x + c1 + c2, dat)
lm_y <- lm(y ~ m2 + m1 + x + c1 + c2, dat)
out <- lm2ptable(list(lm_m1, lm_m2, lm_y))
out$est
#>    lhs op rhs         est
#> 1   m1  ~   x  0.82244298
#> 2   m1  ~  c1  0.17148087
#> 3   m1  ~  c2 -0.18886074
#> 4   m1 ~1     10.81566338
#> 5   m2  ~  m1  0.42078353
#> 6   m2  ~   x -0.11609882
#> 7   m2  ~  c1  0.27753061
#> 8   m2  ~  c2 -0.16194767
#> 9   m2 ~1      3.51937327
#> 10   y  ~  m2  0.52077184
#> 11   y  ~  m1 -0.43533604
#> 12   y  ~   x  0.49285091
#> 13   y  ~  c1  0.09883640
#> 14   y  ~  c2 -0.09604136
#> 15   y ~1      9.46787586